pyproject.toml
[tool.poetry]
name = "rusha_django"
version = "0.1.0"
description = ""
authors = ["Emmanuel <emmanuel.davidson@thejitu.com>"]

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.14.0"
django-mssql-backend = "^2.8.1"
django-mssql = "^1.8"
mssql-django = "^1.1.3"
django-cors-headers = "^3.13.0"
PyYAML = "^6.0"
ansible = "^6.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
